arch: Remove the duplicated syscall.h in each arch
